pyproject.toml
[project]
name = "shrunkiq"
version = "0.1.0"
description = "ShrunkIQ Project"
authors = [
    {name = "Chingis Owana", email = "chingisoinar@gmail.com"},
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9,<3.12"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
    "langchain==0.3.25",
    "langchain-openai==0.3.16",
    "numpy>=1.24.4",
    "openai>=1.78.1",
    "pdf2image>=1.17.0",
    "pymupdf>=1.25.5",
    "opencv-python-headless>=4.11.0.86",
    "langchain-core==0.3.60",
    "tesseract>=0.1.3",
    "pytesseract>=0.3.13",
    "jupyter>=1.1.1",
    "dotenv>=0.9.9",
    "bert-score>=0.3.13",
    "transformers>=4.51.3",
    "hydra-core>=1.3.2",
    "pre-commit==3.8.0",
]

[project.scripts]
shrunkiq = "shrunkiq.cli:main"

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
